Implement a director to manage the plots.
The class myPlotWidget.plotDirector now takes care of linking the various
myPlotWidget to their respective curves. It also handles the buttons that load
the datafiles (e.g. in fileStore) and let the various curves load these data.

The class myPlotWidget.Curve is the prototype of the type of curve expected by
myPlotWidget.myPlotWidget